Update Binary.c
Q1. Consider the problem of adding two n-bit binary integer, stored in two n-element arrays A and B. The sum of the two integers should be stored in binary form in an (n+1)-element array C. Write a C program for the above problem. The input should be two integers in decimal form.